# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SAAS API",
        "description": "API documentation for SAAS Dashboard",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/create_organization": {
            "post": {
                "tags": [
                    "Organizations"
                ],
                "summary": "إنشاء منظمة جديدة",
                "operationId": "c9b57245d469825cd739e417b4de6085",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "phone",
                                    "email",
                                    "type"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "مدرسة النور"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+201234567890"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "info@school.com"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "مدير"
                                    },
                                    "is_master": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "تم إنشاء المنظمة بنجاح",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "تم إنشاء المنظمة بنجاح"
                                        },
                                        "organization": {
                                            "$ref": "#/components/schemas/OrganizationResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/get_organizations": {
            "get": {
                "tags": [
                    "Organizations"
                ],
                "summary": "جلب كل المنظمات",
                "operationId": "29c12a92b2e11dc0664ddab3511d2c5e",
                "responses": {
                    "200": {
                        "description": "تم جلب المنظمات بنجاح",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "تم جلب المنظمات بنجاح"
                                        },
                                        "organization": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/OrganizationResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "OrganizationResource": {
                "title": "OrganizationEmployeeResource",
                "required": [
                    "id",
                    "name",
                    "email",
                    "phone",
                    "type"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "أحمد المعلم"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "teacher@example.com"
                    },
                    "phone": {
                        "type": "string",
                        "example": "+201234567890"
                    },
                    "type": {
                        "type": "string",
                        "example": "معلم"
                    },
                    "is_master": {
                        "type": "integer",
                        "example": 2
                    },
                    "admin_id": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Use bearer token to access this API",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}